+++ tests/heroes/models.py
@@ -1,12 +1,12 @@
 from django.db import models
-from rest_framework_api_key.models import BaseAPIKey
+from rest_framework_api_key.models import AbstractAPIKey
 
 
 class Hero(models.Model):
     objects = models.Manager()
 
 
-class HeroAPIKey(BaseAPIKey):
+class HeroAPIKey(AbstractAPIKey):
     hero = models.OneToOneField(
         Hero, on_delete=models.CASCADE, related_name="api_key"
     )